refactor(wow-spring-boot-starter): remove duplicate EnabledCapable interface
- Remove EnabledCapable interface from wow-spring-boot-starter
- Use EnabledCapable interface from wow-api-naming instead
- Update import statements in multiple files to use the new interface
