+++ wow-spring-boot-starter/src/main/kotlin/me/ahoo/wow/spring/boot/starter/openapi/OpenAPIProperties.kt
@@ -14,7 +14,7 @@
 package me.ahoo.wow.spring.boot.starter.openapi
 
 import me.ahoo.wow.api.Wow
-import me.ahoo.wow.spring.boot.starter.EnabledCapable
+import me.ahoo.wow.api.naming.EnabledCapable
 import org.springframework.boot.context.properties.ConfigurationProperties
 import org.springframework.boot.context.properties.bind.DefaultValue
